Fix PaymentMethodGroupIDs JSON unmarshaling type assertion
JSON numbers are unmarshaled as float64 in Go, not int8. This fixes
the type assertion to correctly handle the float64 type and convert
it to int8.

+++ backoffice/types.go
@@ -368,8 +368,8 @@ func (g *PaymentMethodGroupIDs) UnmarshalJSON(b []byte) error {
 
 	if slice, ok := v.([]any); ok {
 		for _, v := range slice {
-			if id, ok := v.(int8); ok {
-				*g = append(*g, id)
+			if id, ok := v.(float64); ok {
+				*g = append(*g, int8(id))
 			}
 		}
 		return nil
